# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Levels Derm Clinic API",
    "version": "1.0.0",
    "description": "API for cosmetic clinic management system with Arabic support"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/dashboard/kpis": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get dashboard KPIs",
        "description": "Get key performance indicators for the dashboard",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم جلب مؤشرات الأداء بنجاح"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total_patients": {
                          "type": "integer",
                          "example": 145
                        },
                        "total_doctors": {
                          "type": "integer",
                          "example": 19
                        },
                        "today_appointments": {
                          "type": "integer",
                          "example": 17
                        },
                        "today_payments": {
                          "type": "number",
                          "format": "float",
                          "example": 3750.00
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/recent-appointments": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get recent appointments",
        "description": "Get list of recent appointments for the dashboard",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of appointments to retrieve",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم جلب المواعيد الأخيرة بنجاح"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "patient_name": {
                            "type": "string",
                            "example": "نور أحمد محمد"
                          },
                          "doctor_name": {
                            "type": "string",
                            "example": "د. سارة أحمد"
                          },
                          "appointment_date": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-09-17"
                          },
                          "appointment_time": {
                            "type": "string",
                            "example": "12:00"
                          },
                          "status": {
                            "type": "string",
                            "example": "تم التنفيذ"
                          },
                          "payment_method": {
                            "type": "string",
                            "example": "نقدي"
                          },
                          "total_amount": {
                            "type": "number",
                            "format": "float",
                            "example": 500.00
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/today-appointments": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get today's appointments",
        "description": "Get list of today's appointments with patient, doctor, time, status, and payment method",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم جلب مواعيد اليوم بنجاح"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "patient_name": {
                            "type": "string",
                            "example": "نور أحمد"
                          },
                          "doctor_name": {
                            "type": "string",
                            "example": "د. محمد حسن"
                          },
                          "appointment_time": {
                            "type": "string",
                            "example": "12:00"
                          },
                          "status": {
                            "type": "string",
                            "example": "تم التنفيذ"
                          },
                          "payment_method": {
                            "type": "string",
                            "example": "نقدي"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}